new one

+++ backend/controllers/ReceptionController.php
@@ -6,6 +6,7 @@ use backend\models\Comment;
 use backend\models\Note;
 use backend\models\ReceptionTech;
 use backend\models\Students;
+use backend\models\StudentsInfo;
 use Yii;
 use backend\models\Reception;
 use backend\models\ReceptionSearch;
@@ -13,6 +14,8 @@ use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\web\UploadedFile;
+use webvimark\modules\UserManagement\models\User;
+
 /**
  * ReceptionController implements the CRUD actions for Reception model.
  */
@@ -56,20 +59,24 @@ class ReceptionController extends Controller
      */
     public function actionIndex()
     {
-        $form = new Reception();
-        $searchModel = new ReceptionSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
-        $session = Yii::$app->session;
-        $session->set('type_of_reg',2);
-        if (Yii::$app->request->isAjax && $form->load(Yii::$app->request->post())) {
-            Yii::$app->response->format = \yii\web\Response::FORMAT_JSON;
-            return \yii\widgets\ActiveForm::validate($form);
+        if (User::hasRole('Administration')) {
+            $form = new Reception();
+            $searchModel = new ReceptionSearch();
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+            $session = Yii::$app->session;
+            $session->set('type_of_reg', 2);
+            if (Yii::$app->request->isAjax && $form->load(Yii::$app->request->post())) {
+                Yii::$app->response->format = \yii\web\Response::FORMAT_JSON;
+                return \yii\widgets\ActiveForm::validate($form);
+            }
+            return $this->render('index', [
+                'form' => $form,
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        } else {
+            throw new NotFoundHttpException(Yii::t('main','Murojaat huquqi mavjud emas'));
         }
-        return $this->render('index', [
-            'form' => $form,
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
     }
 
     /**
@@ -103,11 +110,11 @@ class ReceptionController extends Controller
         }
         if ($model->load(Yii::$app->request->post())) {
             $model->instance_id = 1;
-            if($model->save(false)) {
-                if(Yii::$app->session->get('type_of_reg')==1){
+            if ($model->save(false)) {
+                if (Yii::$app->session->get('type_of_reg') == 1) {
                     Yii::$app->session->remove('type_of_reg');
                     return $this->redirect(['call-center']);
-                }else{
+                } else {
                     Yii::$app->session->remove('type_of_reg');
                     return $this->redirect(['index']);
                 }
@@ -157,9 +164,13 @@ class ReceptionController extends Controller
      */
     public function actionDelete($id)
     {
+        if (User::hasRole('admin')) {
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
+        }else {
+            throw new NotFoundHttpException(Yii::t('main','Murojaat huquqi mavjud emas'));
+        }
     }
 
     /**
@@ -204,7 +215,7 @@ class ReceptionController extends Controller
     {
         $model = Reception::findOne($changeId);
         $comment = new Comment();
-        $tech = ReceptionTech::find()->where(['reception_id'=>$model->id,'member_id'])->one();
+        $tech = ReceptionTech::find()->where(['reception_id' => $model->id, 'member_id'])->one();
         if (Yii::$app->request->isAjax && $model->load(Yii::$app->request->post())) {
             Yii::$app->response->format = \yii\web\Response::FORMAT_JSON;
             return \yii\widgets\ActiveForm::validate($model);
@@ -230,17 +241,19 @@ class ReceptionController extends Controller
     // Ro'yxatdan o'tish uchun
     public function actionRegister($id = null)
     {
-        $model = new Students();
-        $reception = Reception::findOne($id);
-        if (Yii::$app->request->isAjax && $model->load(Yii::$app->request->post())) {
-            Yii::$app->response->format = \yii\web\Response::FORMAT_JSON;
-            return \yii\widgets\ActiveForm::validate($model);
-        }
-        if ($model->load(Yii::$app->request->post()) && $model->validate()) {
+        if (User::hasRole('Administration')) {
+            $model = new Students();
+            $reception = Reception::findOne($id);
+            $students_info = new StudentsInfo();
+            if (Yii::$app->request->isAjax && $model->load(Yii::$app->request->post())) {
+                Yii::$app->response->format = \yii\web\Response::FORMAT_JSON;
+                return \yii\widgets\ActiveForm::validate($model);
+            }
+            if ($model->load(Yii::$app->request->post())) {
 
                 //img
                 $model->image = UploadedFile::getInstance($model, 'image');
-                if($model->image) {
+                if ($model->image) {
                     $filename = md5(time() . Yii::$app->user->id . $model->image->baseName . rand(1, 1000000) . rand(1, 1000000)) . '.' . $model->image->extension;
 
                     $model->image->saveAs('uploads/img/' . $filename);
@@ -249,49 +262,61 @@ class ReceptionController extends Controller
                 }
                 //pass_file
                 $model->pass_file = UploadedFile::getInstance($model, 'pass_file');
-                if($model->pass_file) {
+                if ($model->pass_file) {
                     $filename = md5(time() . Yii::$app->user->id . $model->pass_file->baseName . rand(1, 1000000) . rand(1, 1000000)) . '.' . $model->pass_file->extension;
                     $model->pass_file->saveAs('uploads/pass/' . $filename);
-                    $model->pass_file = 'uploads/pass/' .$filename;
+                    $model->pass_file = 'uploads/pass/' . $filename;
                 }
                 //file
                 $model->file = UploadedFile::getInstance($model, 'file');
-                if($model->file) {
+                if ($model->file) {
                     $filename = md5(time() . Yii::$app->user->id . $model->file->baseName . rand(1, 1000000) . rand(1, 1000000)) . '.' . $model->file->extension;
                     $model->file->saveAs('uploads/file/' . $filename);
                     $model->file = 'uploads/file/' . $filename;
                 }
 
                 $model->active = 1;
-                $reception->instance_id = 4;
-                $reception->save();
-            if ($model->save()){
-                return $this->redirect(Yii::$app->request->referrer);
+                if ($model->save()) {
+                    $reception->instance_id = 4;
+                    $reception->save(false);
+                    $students_info->students_id = $model->id;
+                    $students_info->save(false);
+                    \Yii::$app->session->setFlash('success', Yii::t('main', 'Student Reg'));
+                    return $this->redirect(['index']);
+                }
             }
-        }
 
-        return $this->renderAjax('students', [
-            'model' => $model,
-            'reception' => $reception
-        ]);
+            return $this->render('students', [
+                'model' => $model,
+                'reception' => $reception,
+                'students_info' => $students_info,
+            ]);
+        } else {
+            throw new NotFoundHttpException(Yii::t('main','Murojaat huquqi mavjud emas'));
+        }
     }
 
     // class - center
-    public function actionCallCenter(){
-        $form = new Reception();
-        $searchModel = new ReceptionSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
-        $session = Yii::$app->session;
-        $session->set('type_of_reg',1);
-        if (Yii::$app->request->isAjax && $form->load(Yii::$app->request->post())) {
-            Yii::$app->response->format = \yii\web\Response::FORMAT_JSON;
-            return \yii\widgets\ActiveForm::validate($form);
+    public function actionCallCenter()
+    {
+        if (User::hasRole('call-center')) {
+            $form = new Reception();
+            $searchModel = new ReceptionSearch();
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+            $session = Yii::$app->session;
+            $session->set('type_of_reg', 1);
+            if (Yii::$app->request->isAjax && $form->load(Yii::$app->request->post())) {
+                Yii::$app->response->format = \yii\web\Response::FORMAT_JSON;
+                return \yii\widgets\ActiveForm::validate($form);
+            }
+            return $this->render('call-center', [
+                'form' => $form,
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        } else {
+            throw new NotFoundHttpException('#404.');
         }
-        return $this->render('call-center', [
-            'form' => $form,
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
     }
 }
